requirements.txt
python-dotenv==1.1.1
aiogram==3.22.0